Refactor assets.js into multiple files.
