--- src/routes/assets.js
@@ -1,303 +1,0 @@
-'use strict';
-// Handler functions for the /assets endpoint.
-
-const async = require('async');
-const fs = require('fs');
-const path = require('path');
-const request = require('request');
-const crypto = require('crypto');
-const urljoin = require('urljoin');
-const _ = require('lodash');
-const config = require('../config');
-const storage = require('../storage');
-const log = require('../logging').getLogger();
-
-/**
- * @description Calculate a checksum of an uploaded file's contents to generate
- *   the fingerprinted asset name.
- */
-function fingerprintAsset (asset, callback) {
-  var sha256sum = crypto.createHash('sha256');
-  var assetFile = fs.createReadStream(asset.path);
-  var chunks = [];
-
-  assetFile.on('data', function (chunk) {
-    sha256sum.update(chunk);
-    chunks.push(chunk);
-  });
-
-  assetFile.on('error', callback);
-
-  assetFile.on('end', function () {
-    var digest = sha256sum.digest('hex');
-    var ext = path.extname(asset.name);
-    var basename = path.basename(asset.name, ext);
-    var fingerprinted = basename + '-' + digest + ext;
-
-    log.debug({
-      action: 'assetstore',
-      originalAssetName: asset.name,
-      assetFilename: fingerprinted,
-      assetContentType: asset.type,
-      message: 'Asset fingerprinted successfully.'
-    });
-
-    callback(null, {
-      key: asset.key,
-      original: asset.name,
-      chunks: chunks,
-      filename: fingerprinted,
-      type: asset.type
-    });
-  });
-}
-
-/**
- * @description Upload an asset's contents to the asset container.
- */
-function publishAsset (asset, callback) {
-  storage.storeAsset(asset, function (err, asset) {
-    if (err) {
-      return callback(err);
-    }
-
-    log.debug({
-      action: 'assetstore',
-      assetFilename: asset.filename,
-      assetContentType: asset.type,
-      message: 'Asset uploaded successfully.'
-    });
-
-    callback(null, asset);
-  });
-}
-
-/**
- * @description Give this asset a name. The final name and CDL URI of this
- *   asset will be included in all outgoing metadata envelopes, for use by
- *   layouts.
- */
-function nameAsset (asset, callback) {
-  storage.nameAsset(asset, function (err, asset) {
-    if (err) {
-      return callback(err);
-    }
-
-    log.debug({
-      action: 'assetstore',
-      originalAssetFilename: asset.original,
-      assetName: asset.key,
-      message: 'Asset named successfully.'
-    });
-
-    callback(null, asset);
-  });
-}
-
-/**
- * @description Create and return a function that processes a single asset.
- */
-function makeAssetHandler (shouldName) {
-  return function (asset, callback) {
-    log.debug({
-      action: 'assetstore',
-      originalAssetName: asset.name,
-      message: 'Asset upload request received.'
-    });
-
-    var steps = [
-      async.apply(fingerprintAsset, asset),
-      publishAsset
-    ];
-
-    if (shouldName) {
-      steps.push(nameAsset);
-    }
-
-    async.waterfall(steps, callback);
-  };
-}
-
-/**
- * @description Enumerate all named assets.
- */
-var enumerateNamed = exports.enumerateNamed = function (callback) {
-  storage.findNamedAssets(function (err, assetVars) {
-    if (err) {
-      return callback(err);
-    }
-
-    var assets = {};
-
-    for (var i = 0; i < assetVars.length; i++) {
-      var assetVar = assetVars[i];
-      assets[assetVar.key] = assetVar.publicURL;
-    }
-
-    callback(null, assets);
-  });
-};
-
-/**
- * @description Append a list of asset names to a message before it's logged.
- */
-function withAssetList (message, originalAssetNames) {
-  var subset = originalAssetNames.slice(0, 3);
-
-  if (originalAssetNames.length > 3) {
-    subset.push('..');
-  }
-
-  var joined = subset.join(', ');
-
-  return message + ': ' + joined + '.';
-}
-
-/**
- * @description Fingerprint and upload static, binary assets to the
- *   CDN-enabled ASSET_CONTAINER. Return a JSON object containing a
- *   map of the provided filenames to their final, public URLs.
- */
-exports.accept = function (req, res, next) {
-  var originalAssetNames = [];
-  var assetData = Object.getOwnPropertyNames(req.files).map(function (key) {
-    var asset = req.files[key];
-    asset.key = key;
-    originalAssetNames.push(key);
-    return asset;
-  });
-
-  log.debug({
-    action: 'assetstore',
-    apikeyName: req.apikeyName,
-    assetCount: assetData.length,
-    originalAssetNames: originalAssetNames,
-    message: withAssetList('Asset upload request received', originalAssetNames)
-  });
-
-  var reqStart = Date.now();
-
-  async.map(assetData, makeAssetHandler(req.query.named), function (err, results) {
-    if (err) {
-      var statusCode = err.statusCode || 500;
-
-      log.error({
-        action: 'assetstore',
-        statusCode: statusCode,
-        apikeyName: req.apikeyName,
-        error: err.message,
-        stack: err.stack,
-        message: withAssetList('Unable to upload one or more assets', originalAssetNames)
-      });
-
-      res.send(statusCode, {
-        apikeyName: req.apikeyName,
-        error: 'Unable to upload one or more assets.'
-      });
-
-      return next(err);
-    }
-
-    var summary = {};
-    results.forEach(function (result) {
-      summary[result.original] = result.publicURL;
-    });
-    log.info({
-      action: 'assetstore',
-      statusCode: 200,
-      apikeyName: req.apikeyName,
-      totalReqDuration: Date.now() - reqStart,
-      message: withAssetList('All assets have been uploaded successfully', originalAssetNames)
-    });
-
-    res.send(summary);
-    next();
-  });
-};
-
-exports.list = function (req, res, next) {
-  log.debug('Asset list requested.');
-
-  enumerateNamed(function (err, assets) {
-    if (err) {
-      log.error({
-        action: 'assetlist',
-        statusCode: err.statusCode || 500,
-        message: 'Unable to list assets.',
-        error: err.message,
-        stack: err.stack
-      });
-      return next(err);
-    }
-
-    if (config.proxyUpstream()) {
-      let url = urljoin(config.proxyUpstream(), 'assets');
-      log.debug('Requesting upstream assets', {
-        upstreamURL: url
-      });
-
-      request({ url, json: true }, (err, response, body) => {
-        if (err || response.statusCode !== 200) {
-          log.error('Unable to retrieve upstream assets.', {
-            action: 'assetlist',
-            statusCode: response.statusCode,
-            error: err ? err.message : null,
-            stack: err ? err.stack : null
-          });
-
-          let e = new Error('Unable to retrieve upstream assets.');
-          e.statusCode = 502;
-          return next(e);
-        }
-
-        assets = _.merge(body, assets);
-
-        res.send(assets);
-        next();
-      });
-      return;
-    }
-
-    res.send(assets);
-    next();
-  });
-};
-
-exports.retrieve = function (req, res, next) {
-  log.debug({
-    action: 'assetretrieve',
-    message: 'Asset requested directly.'
-  });
-
-  var reqStart = Date.now();
-
-  storage.getAsset(req.params.id, function (err, asset) {
-    if (err) {
-      log.error({
-        action: 'assetretrieve',
-        statusCode: err.statusCode || 500,
-        message: 'Unable to retrieve asset',
-        error: err.message,
-        stack: err.stack
-      });
-      return next(err);
-    }
-
-    // Bypass restify's formatter to keep it from being "helpful"
-    res.setHeader('content-type', asset.contentType);
-    res.writeHead(200);
-    res.write(asset.body);
-    res.end();
-
-    log.info({
-      action: 'assetretrieve',
-      statusCode: 200,
-      assetFilename: req.params.id,
-      assetContentType: asset.contentType,
-      totalReqDuration: Date.now() - reqStart,
-      message: 'Asset request successful.'
-    });
-
-    next();
-  });
-};
