+++ src/routes/assets/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+exports.store = require('./store').handler;
+exports.retrieve = require('./retrieve').handler;
+exports.list = require('./list').handler;
